// src/lib/types.ts
export interface HtmlSpanProps extends StrictHtmlSpanProps {
  [key: string]: any;
}

export interface StrictHtmlSpanProps {
  children?: React.ReactNode;
}

// ======================================================
// Types
// ======================================================

export type SemanticShorthandItemFunc<TProps> = (component: React.ReactType<TProps>, props: TProps, children?: React.ReactNode | React.ReactNodeArray) => React.ReactElement<any> | null;
export type SemanticShorthandContent = React.ReactNode;
export type SemanticShorthandItem<TProps> = React.ReactNode | TProps | SemanticShorthandItemFunc<TProps>;
// ...
export interface FCX<P> extends React.FC<P> {

  create: Function;
}